Move TruncateToEthAddress to pallet file, update licence
